+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 * server: make it possible to override instanceId/prefix in RequestIdGenerator to support vendor-specific instance id headers, like FLY_ALLOC_ID
+* server: metrics() now exposes RSS memory usage of the whole JVM process on Linux
 * json: report line numbers in JsonParseException for better debugging of json parsing errors
 * json: make parser stricter according to the spec: disallow newlines inside of strings and unary plus
 * json: increase JsonRenderer performance by ~10%